+++ wp-content/plugins/advisories-headers/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Only supply a CSP on the front-end of the site
 * Remove 'unsafe-inline' from script-src in CSP and add a nonce to all scripts
 * Remove 'unsafe-inline' from style-src in CSP and add a nonce to all scripts
+* Remove 'contain-intrinsic-size' provided by WordPress Core
+* Remove some enqueued styles from WordPress Core where we cannot add a nonce
 
 ## [v1.0.1] - 2025-06-26